include key-id value in the strongbox prefix
Helps identify the key you need to decrypt the file, in case the secrets
were not rotated when the key on disk changed.

Also paves the way to fix the bug where cloning files alphabetically
preceding `.stringbox-keyid` fails do to absence of the key on disk. We
can now use the key-id from the header instead.